+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.11.29
 
+* added `Predicate.is_instance`